+++ web-wallet/src/lib/stores/stores.d.ts
@@ -144,7 +144,7 @@ type WalletStoreServices = {
     gas: import("$lib/vendor/w3sper.js/src/mod").Gas
   ) => Promise<TransactionInfo>;
 
-  withdrawReward: (
+  claimRewards: (
     amount: bigint,
     gas: import("$lib/vendor/w3sper.js/src/mod").Gas
   ) => Promise<TransactionInfo>;