+++ web-wallet/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update Transactions list design [#1922]
+- Reword "Withdraw Rewards" operation to "Claim Rewards" [#3076]
 
 ### Removed
 
@@ -48,7 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added gas settings validation on Unstake / Widthdraw Rewards flows [#2000]
+- Added gas settings validation on Unstake / Withdraw Rewards flows [#2000]
 - Add temporary link to the block explorer on the dashboard [#2882]
 - Add Staking-Related Functionality Utilizing w3sper [#3006]
 - Add minimum stake amount supplied by w3sper instead of using an env var [#3010]
@@ -400,6 +401,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#3069]: https://github.com/dusk-network/rusk/issues/3069
 [#3071]: https://github.com/dusk-network/rusk/issues/3071
 [#3073]: https://github.com/dusk-network/rusk/issues/3073
+[#3076]: https://github.com/dusk-network/rusk/issues/3076
 
 <!-- VERSIONS -->